Separate /product/{id} client logic into into sub-components
